Issue 110: Get docker working, including working with safer privileges.
It seems that the group id cannot be passed as a parameter because the `COPY --chown=` command does not support that.
Use a relatively high number, such as 3001 for the uid and gid to reduce potential id conflicts with the system loaded by docker.
Use a single variable ARG declaration for each variable and then import them into each stage.
This must be done because ARG is lost between stages and we really should avoid exposing these via ENV.

Use a custom work directory owned by a created user.
Run maven as that user to avoid doing anything as root (security improvement).
Run the service as that user from within that directory.
The jar is moved and renamed for convenience.

Maven (and possibly java) can be picky about arguments, so use the bracket notation for passing arguments to both maven and java.