Issue 150: Fix problems with escaping and update unit tests.
This primarily focuses solving the problems with Fedora but many of the changes are broader than Fedora such that they may impact DSpace.

The data from the client should throw an error if the data is not properly escaped.
The escaping must not happen here.
This reverts the changes from commit eda72b902cd3ea18660af5180ffd15d96903ffbe.

Simplify the `getRdf()` design.
- Replace `logger.info()` wrapped inside of `logger.isDebugEnabled()` with the simpler `logger.debug()`.
- Extend the `NotFoundException` to add the exception to the message.
- Remove excessive logging.

The unit tests throw `org.apache.jena.riot.RiotException` errors to match what the server should be throwing when there is unencoded spaces.
It is unclear whether or not we should or should not convert this into a `RestClientException`.
This is left alone for further discussion.

The `encodeSpaces()` is entirely removed.
The project should not be encoding bad data.
Instead, bad data should throw an error, such as the `RiotException`.
The DSpace code will need to be reviewed for accuracy and regressions after this change.

Change `fetchImageInfo()` logic to handle `RestClientException` and return `NotFoundException`.

An observed potential problem is:
```java
    private String getResourceId(String url) throws URISyntaxException {
        try {
            id = resourceResolver.lookup(url);
        } catch (NotFoundException e) {
            id = resourceResolver.create(url);
        }
```
This logic will fail on `NotFoundException`.
While this may not be the case here, there are a lot of places that translate the actual errors into `NotFoundException` even when the error is something other than `NotFoundException`.
This could potentially result in an attempt to create data when it should not.
I have not verified this behavior.

Several of the unit test regarding Fedora are refactored.
There are now multiple tests for encoded, decoded, and neither (uncoded).
The decoded tests are designed to throw exceptions that should be thrown under normal operations.

The Dspace unit tests will be updated in a subsequent commit.
Symbolic links are added for the `json` and `rdf` directories as a temporary measure to support the DSpace tests in their current form.