Issue 150: Update unit tests for Dspace and update Fedora tests.
This is a follow up to commit 41731b8dc41b20cbf61fdf0e440e18bb1e933f34.

This primarily focuses on updating the DSpace unit tests to match the new behavior that the Fedora unit tests are now following.
This only applies to the decode, encode, and uncode unit tests.

The symbolic links to the uncoded `rdf` and `json` are remaining because there are other unrelated tests still using these files at this path.
It is suggested to update this in some future commit.

Fix problems observed in the Fedora tests (typos and other such mistakes).